+++ types/index.ts
@@ -1,3 +1,3 @@
 export type { LayoutProps } from "./LayoutProps";
 export type { IconProps } from "./IconProps";
-export type { MenuId } from "./aaa";
+export type { MenuId } from "./MenuId";
